docker/python/requirements.txt
absl-py==1.0.0
anyio==3.5.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
fastapi==0.75.1
flatbuffers==2.0
fonttools==4.31.2
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.3
ipython==8.2.0
jedi==0.18.1
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.2
kiwisolver==1.4.2
kt-legacy==1.0.4
libclang==13.0.0
Markdown==3.3.6
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.34
stack-data==0.2.0
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
wcwidth==0.2.5
Werkzeug==2.1.0
wrapt==1.14.0
zipp==3.7.0
